Crate hahadaxigua

Hi, my name is Aden, a backend programmer in Shanghai/Chain, who had a big interest in technology, digital products.

I have been using Golang as my main programming-language for two years, and Rust is my second favorite language. I’m looking for a backend-development job which using golang and rust.

You can find me on GitHub

You can also send me an email.

Modules

current

Junior Golang backend developer.

education

Bachelor of BNetwork Engineering from 2017 to 2021, in Nantong University

Macros

cargo_temp

A CLI tool that allows you to create a temporary new Rust project using cargo with already installed dependencies. [Repository]

create_process_w

A crate to create and handle processes on Windows (similar too std::process). [Repository]

intro_to_rust

My vision of the Rust Programming Language as an introduction to Rust. [Repository]

tui_do

Yet Another To-Do App in your Terminal. (Under development) [Repository]

yewprint_playground

A Yew and Yewprint playground using wasm-run. [Repository]

Constants

OS

Ubuntu, Darwin.

PROGRAMMING_LANGUAGES

Golang, Rust, Java.

SPOKEN_LANGUAGES

Chinese(5.Native) English (3. Minimum Professional Proficiency).

TECHNOLOGIES

Git, Docker, .

Traits

Autodidact

Enthusiast

Flexible

Passionate

TeamPlayer

WantToLearn